requirements.txt
argcomplete==3.3.0
asgiref==3.8.1
boto3==1.34.88
botocore==1.34.88
certifi==2024.2.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.4
django-environ==0.11.2
django-s3-storage==0.15.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
durationpy==0.6
gunicorn==22.0.0
hjson==3.1.0
idna==3.7
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
packaging==24.0
placebo==0.9.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.2
troposphere==4.7.0
urllib3==2.2.1
Werkzeug==3.0.2
zappa==0.59.0
